pyproject.toml
[project]
name = "run-coach"
version = "0.1.0"
description = "AI-powered running coach CLI"
requires-python = ">=3.11.5"
dependencies = [
    "garminconnect>=0.2.38",
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "langgraph>=1.0.10",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "psycopg[binary]>=3.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "alembic>=1.18.4",
    "google-cloud-storage>=3.9.0",
    "google-auth>=2.49.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
    "types-pyyaml>=6.0.12.20250915",
]
